requirements.txt
git+https://github.com/openai/whisper.git
pytube
torch
farm-haystack
openai
gradio
